Fix compatibility with Python 3.14
Nesting argument groups was deprecated in Python 3.11 and will be
removed in the upcoming Python 3.14:
https://github.com/python/cpython/issues/127133

Trying to add a nested group will trigger a `ValueError`.

To prepare the code for Python 3.14, remove the nested groups and add
all arguments at the top level. The mutual exclusion between `--kill`
and `--foreground`/`--replace` is already checked separately within the
`main()` function in `scripts/powerline-daemon`, so the user-observable
behaviour remains unchanged.