Fix compatibility with Python 3.15
The undocumented function `glob.glob1()` was deprecated in Python 3.13
and removed in Python 3.15.

Replace it with a call to `glob.glob()` with the `root_dir` argument
set. Note that this requires at least Python 3.10, which was released in
October 2021. This is deemed acceptable as Python 3.9 was declared
end-of-life in October 2025.

Remove two unused imports of `glob1`.